Refactor di GarageView
- Creazione della component searchbar per corretto utilizzo in hoomeview e garageview;
- Creazione di composable per corretta gestione dei dati
