+++ client/src/views/HomeView.vue
@@ -1,16 +1,26 @@
 <script setup>
 import { ref } from 'vue'
+import { useRouter } from 'vue-router'
 import Header from '../components/Header.vue'
 import Footer from '../components/Footer.vue'
+import SearchBar from '../components/SearchBar.vue'
 
+const router = useRouter()
 // Logica di ricerca invariata
 const searchLocation = ref('')
 const checkIn = ref('')
 const checkOut = ref('')
 
 const handleSearch = () => {
-  console.log('Ricerca:', searchLocation.value, checkIn.value, checkOut.value)
-  alert(`Cerco parcheggi a ${searchLocation.value} dal ${checkIn.value} al ${checkOut.value}`)
+  // Passiamo i parametri tramite URL alla pagina dei garage
+  router.push({
+    name: 'garage',
+    query: {
+      location: searchLocation.value,
+      checkIn: checkIn.value,
+      checkOut: checkOut.value
+    }
+  })
 }
 </script>
 
@@ -19,7 +29,7 @@ const handleSearch = () => {
     <Header />
 
     <main class="main-content">
-      
+
       <section class="hero-header">
         <div class="hero-text">
           <h1>Sostare non è mai stato così <span class="highlight">semplice</span>.</h1>
@@ -28,32 +38,8 @@ const handleSearch = () => {
       </section>
 
       <section class="search-container">
-        <!-- significa: "Quando l'utente preme Invio, PREVIENI il comportamento normale del browser (che ricaricherebbe la pagina sfasando tutto) ed esegui solo la mia funzione handleSearch"-->
-        <form @submit.prevent="handleSearch" class="search-box">
-          
-          <div class="input-group location-group">
-            <div class="icon">📍</div>
-            <div class="fields">
-              <label>Dove vuoi parcheggiare?</label>
-              <input type="text" v-model="searchLocation" placeholder="Città, indirizzo o stazione..." required>
-            </div>
-          </div>
-
-          <div class="input-group">
-            <label>Check-in</label>
-            <input type="datetime-local" v-model="checkIn" required>
-          </div>
-
-          <div class="input-group">
-            <label>Check-out</label>
-            <input type="datetime-local" v-model="checkOut" required>
-          </div>
-
-          <button type="submit" class="search-btn">
-             Cerca
-          </button>
-          
-        </form>
+        <SearchBar v-model:location="searchLocation" v-model:checkIn="checkIn" v-model:checkOut="checkOut"
+          :showSubmitButton="true" @search="handleSearch" />
       </section>
 
       <section class="why-parkly">
@@ -84,7 +70,6 @@ const handleSearch = () => {
 </template>
 
 <style scoped>
-
 .home-wrapper {
   display: flex;
   flex-direction: column;
@@ -101,7 +86,8 @@ const handleSearch = () => {
 .hero-header {
   background-color: var(--deep-blue);
   color: var(--white);
-  padding: 5rem 2rem 10rem 2rem; /* Spazio extra in basso per il "floating" */
+  padding: 5rem 2rem 10rem 2rem;
+  /* Spazio extra in basso per il "floating" */
   text-align: center;
 }
 
@@ -112,15 +98,17 @@ const handleSearch = () => {
 }
 
 .highlight {
-  color: var(--accent-blue); /* Un azzurro più chiaro per l'enfasi sul testo blu scuro */
+  color: var(--accent-blue);
+  /* Un azzurro più chiaro per l'enfasi sul testo blu scuro */
 }
+
 /* Stile Sottotitolo Hero:
   - Opacità 80%: dà priorità visiva al titolo principale
   - Max 700px + margin auto: centra il testo e previene righe troppo lunghe
 */
 .hero-text p {
   font-size: 1.3rem;
-  color: rgba(255,255,255,0.8);
+  color: rgba(255, 255, 255, 0.8);
   max-width: 700px;
   margin: 0 auto;
 }
@@ -130,94 +118,10 @@ const handleSearch = () => {
   max-width: 1050px;
   margin: 0 auto;
   padding: 0 1rem;
-  margin-top: -4rem; /* Trucco CSS: fluttuazione sopra la zona blu */
+  margin-top: -4rem;
+  /* Trucco CSS: fluttuazione sopra la zona blu */
   position: relative;
   z-index: 10;
-}
-
-.search-box {
-  display: flex;
-  background: var(--white);
-  padding: 1rem;
-  border-radius: 12px;
-  box-shadow: 0 10px 30px rgba(0,0,0,0.1); /* Ombra più morbida per eleganza */
-  gap: 10px;
-}
-
-/* Rielaborazione input group per uno stile pulito "app" */
-.input-group {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  padding: 0.5rem 1rem;
-  background: var(--white);
-  border-radius: 8px;
-  border: 2px solid var(--border-light); /* Bordo grigio chiaro di base */
-  transition: all 0.3s;
-}
-
-/* Effetto focus sul gruppo di input */
-.input-group:focus-within {  
-  border-color: var(--primary-blue);
-  box-shadow: 0 0 0 3px rgba(0, 102, 204, 0.1);
-}
-
-.location-group {
-  flex: 1.5; /* Più spazio per l'indirizzo */
-  flex-direction: row;
-  align-items: center;
-  gap: 10px;
-}
-
-.location-group .fields {
-  display: flex;
-  flex-direction: column;
-  flex: 1;
-}
-
-.location-group .icon {
-  font-size: 1.3rem;
-}
-
-.input-group label {
-  font-size: 0.8rem;
-  font-weight: bold;
-  color: var(--primary-blue); /* Label colorate di blu */
-  margin-bottom: 4px;
-}
-
-.input-group input {
-  border: none;
-  outline: none;
-  font-size: 1rem;
-  font-family: inherit;
-  width: 100%;
-  color: var(--text-dark);
-}
-
-.input-group input::placeholder {
-  color: #999;
-}
-
-/* PULSANTE BLU ACCESO (Call to Action) */
-.search-btn {
-  background-color: var(--primary-blue);
-  color: white;
-  border: none;
-  font-size: 1.2rem;
-  font-weight: bold;
-  padding: 0 2.5rem;
-  border-radius: 8px;
-  cursor: pointer;
-  transition: all 0.3s ease;
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-.search-btn:hover {
-  background-color: #0052a3; /* Un blu leggermente più scuro al passaggio */
-  transform: translateY(-2px);
 }
 
 /* 4. SEZIONE VANTAGGI SOTTOSTANTE */
@@ -245,7 +149,7 @@ const handleSearch = () => {
   padding: 2.5rem 1.5rem;
   text-align: center;
   border-radius: 12px;
-  box-shadow: 0 4px 15px rgba(0,0,0,0.03);
+  box-shadow: 0 4px 15px rgba(0, 0, 0, 0.03);
   border: 1px solid var(--border-light);
   transition: transform 0.3s;
 }
@@ -273,21 +177,11 @@ const handleSearch = () => {
 
 /* Responsività per Smartphone (Ancora più pulita) */
 @media (max-width: 900px) {
-  .search-box {
-    flex-direction: column;
-    padding: 0.5rem;
-    gap: 5px;
-  }
-  .input-group {
-    border-radius: 6px;
-  }
-  .search-btn {
-    padding: 1.2rem;
-    justify-content: center;
-  }
+
   .cards-grid {
     flex-direction: column;
   }
+
   .hero-text h1 {
     font-size: 2.2rem;
   }
